reorganize to facilitate fixture reuse
