+++ tests/models/independent/test_ProtocolInterface.py
@@ -9,9 +9,29 @@ __email__ = "vreuter@virginia.edu"
 
 
 
+@pytest.mark.skip("Not implemented")
 class PipelinePathResolutionTests:
     """ Project requests pipeline information via an interface key. """
 
 
-    def test_pipeline_interface_path(self):
+    def test_no_path(self, piface_config_bundles):
         pass
+
+
+    def test_relative_path(self, piface_config_bundles):
+        pass
+
+
+    def test_absolute_path(self, piface_config_bundles):
+        pass
+
+
+    def test_pipeline_interface_path(self, piface_config_bundles):
+        pass
+
+
+
+@pytest.mark.skip("Not implemented")
+class ProtocolInterfacePipelineSampleSubtypeTests:
+    """ ProtocolInterface attempts import of pipeline-specific Sample. """
+    pass
